Allow float values for the initial position of other players

+++ game/network.py
@@ -92,8 +92,8 @@ class NewPlayer(Command):
     """
 
     arguments = [('identifier', Integer()),
-                 ('x', Integer()),
-                 ('y', Integer()),
+                 ('x', Float()),
+                 ('y', Float()),
                  ('speed', Integer())]
